Fix rendering of container, remember story
Make it so the header renders without margin, and the body is
scrollable, rather than the whole page being scrollable to keep the
header visible.

Use async storage so we remember which story we were on while
developing. It was rather annoying when you'd update a story and then
have to scroll and click the story again to see changes.

Also add build command so we can deploy web to Netlify.