Add multiselect as Field type (#41)
* Fix rendering of container, remember story

Make it so the header renders without margin, and the body is
scrollable, rather than the whole page being scrollable to keep the
header visible.

Use async storage so we remember which story we were on while
developing. It was rather annoying when you'd update a story and then
have to scroll and click the story again to see changes.

Also add build command so we can deploy web to Netlify.

* Fix stories using hooks

We need to encapsulate the hooks in a component rather than just a
story function, otherwise React gets mad about different numbers
of hooks. Annoying but works.

* Update TapToEdit component

Use Field for the editing, giving TapToEdit support for any of the types
that Field supports. Also use hooks to contain the values better in
TapToEdit without requiring it to be a controlled component.

* Add multiselect as Field type

This allows selecting multiple options for a select list. They options
are rendered as Switches to toggle on and off. Also add support to
TapToEdit for multiselect.

* Add expo cli so netlify can build the app

* Upgrade ferns-ui in demo