Use react-native-calendar for web date picker
I tried a few different date picker libraries, but most want to control
the input field as well, which doesn't work well with TextField. Show a
calendar below the field when tapping a TextField of type date on web.
In the future, this should be a popover or use the ActionSheet so it
doesn't mess with the layout of the form containing the TextField.

This package comes from Wix, who usually make high quality RN software.