Change TapToEdit from uncontrolled to controlled component (#97)
Previously, TapToEdit was managing its state internally only. This
caused a bunch of data sync issues (and was more useful in the pre-hooks
era). Switch this to a controlled component to fix up a few bugs it has
been experiencing.

Also update package.json for a few missing peer deps.