Add Table component (#205)
* Add Table component

Add Table and components for building out tables using rows, headers,
and header cells. Use TableContext to more easily coordinate between all
the different components in a Table.

One choice is to require widths for the table on setup (not based on the
content) and use horizontal scrolling to put tables at full size on
mobile devices. This might not always be the right choice, but tables
are always hard to do with mobile, so this at least gives a consistent
looking table on all three platforms.

* Remove errant date time removal

* Respond to review

* Fix table width when expandable content is present