Create new date modal, update modal for mobile (#221)
* Create new date modal, update modal for mobile

For TextField and Field, when set to date, show a modal with a calendar
in it to pick the date. Add a time and datetime type to both, showing a
time picker (either spinner on mobile or text inputs on web) and
calendar + time picker respectively. This solves the various z-index
issues, styling issues, and issues trying to pick dates in constrained
spaces.

On mobile, show an action sheet whenever we would show a modal on web.
This mimics the iOS standard for showing sheets. In the future, we
should make action sheet behave more like a modal on Android (similar to
its normal sheets)

* Fix up error where 12:00 wasn't a valid time

* Fix time UI on web, fix a couple time bugs