Segment and Modal fixes (#323)
* Fix segmented control overflow

If one of the segmented control titles wound up as two lines,
it would align left instead of center. The lack of padding looked
weird too.

* Fix modal header

The close button was getting its own row, pushing the header down,
causing the top to look very weird on web. On mobile, fix the primary
button to align better with the title and close.

* lint

* lint

* Remove ui-demo expo doctor

I think something with our linking is breaking this. We still run doctor
on UI, so hopefully that catches actual problems.