Fix DateTimeField, add tests (#854)
* Fix DateTimeField, add tests

Add header and error to DateTimeField
Fix DateTimeField type="date" rendering the previous day in the text
field
Add tests for DateTimeField

* Fix up typing dates

There was an issue with state not getting set correctly.

* Change the handling of onChange and onBlur

Use a ref to keep the pending values for onBlur. The problem was
onBlur was running before onChange had fully completed, so it was
blurring the value from one state change ago. Use a ref to keep data in
sync and simplify some handling of when we call onChange.

* Respond to review